globals: caids in struct s_cardsystem must be 16 bit.
git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@6629 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ globals.h
@@ -724,7 +724,7 @@ struct s_cardsystem {
 	void			(*post_process)();
 	int32_t		(*get_emm_type)();
 	void			(*get_emm_filter)();
-	uchar			caids[2];
+	uint16_t		caids[2];
 };
 
 #ifdef IRDETO_GUESSING
